rename dump-ir to print-ir. combine with print-js
